+++ src/components/blog/Poososo/PostCardVariants.ts
